+++ RELEASE-NOTES.md
@@ -22,6 +22,7 @@
 * #45 DataFrame.renameColumns(UnaryOperator<String>) to upper/lower case column names
 * #46 Move interfaces with public static methods for creating lambdas to the root package
 * #47 Series.map 
+* #48 dflib-jdbc: saver must support Month, Year, enums, etc
 
 ## Release 0.5